move sorter into separate component Sorter, add condition to show Sorter on Menu page
