+++ client/src/components/ListItems/ListItem.js
@@ -5,56 +5,21 @@ import PropTypes from 'prop-types';
 import Grid from '@mui/material/Grid';
 import Divider from '@mui/material/Divider';
 import { Box, MenuItem, TextField } from '@mui/material';
+import { useLocation } from 'react-router-dom';
 import { gridStylesItemPartners, gridStylesItemProducts, gridStylesContainer, stylesSortSelect } from './styles';
 import AppPagination from '../Pagination/Pagination';
 import useBreakpoint from '../../customHooks/useBreakpoint';
+import Sorter from '../Sorter/Sorter';
 
 const ListItems = ({ title, items, itemComponent, actions, pagination, anchor, type }) => {
-  let currencies;
-  if (type === 'partners') {
-    currencies = [
-      {
-        value: 'Rating UP',
-        label: 'Rating UP',
-      },
-      {
-        value: 'Rating DOWN',
-        label: 'Rating DOWN',
-      },
-      {
-        value: 'Default',
-        label: 'Default',
-      },
-    ];
-  } else {
-    currencies = [
-      {
-        value: 'Price UP',
-        label: 'Price UP',
-      },
-      {
-        value: 'Price DOWN',
-        label: 'Price DOWN',
-      },
-      {
-        value: 'Rating UP',
-        label: 'Rating UP',
-      },
-      {
-        value: 'Rating DOWN',
-        label: 'Rating DOWN',
-      },
-      {
-        value: 'Default',
-        label: 'Default',
-      },
-    ];
-  }
+  const { pathname } = useLocation();
 
   const [selectedValueSortBy, setSelectedValueSortBy] = React.useState('');
+
   useEffect(() => {
     setSelectedValueSortBy('Default');
   }, [items]);
+
   const itemsCopy = React.useMemo(() => {
     const copy = [...items];
     if (selectedValueSortBy === 'Price UP') {
@@ -71,9 +36,6 @@ const ListItems = ({ title, items, itemComponent, actions, pagination, anchor, t
     }
     return copy;
   }, [items, selectedValueSortBy]);
-  const handleSelectChangeSortBy = (event) => {
-    setSelectedValueSortBy(event.target.value);
-  };
 
   const breakpoint = useBreakpoint();
   const productsPerPageMap = {
@@ -120,35 +82,24 @@ const ListItems = ({ title, items, itemComponent, actions, pagination, anchor, t
       >
         {title}
       </Typography>
-      <Box sx={{ width: '100%', height: '40px', mb: '40px', paddingRight: '30px', textAlign: 'end' }}>
-        <TextField
-          sx={stylesSortSelect}
-          id="standard-select-currency"
-          size="small"
-          select
-          label="Sort by"
-          defaultValue="Default"
-          variant="standard"
-          value={selectedValueSortBy}
-          onChange={handleSelectChangeSortBy}
-        >
-          {currencies.map((option) => (
-            <MenuItem key={option.value} value={option.value}>
-              {option.label}
-            </MenuItem>
-          ))}
-        </TextField>
-      </Box>
+
+      { pathname === '/menu' && (
+      <Sorter
+        type={type}
+        selectedValueSortBy={selectedValueSortBy}
+        setSelectedValueSortBy={setSelectedValueSortBy}
+      />
+      )}
+
       <Grid container spacing={0} sx={gridStylesContainer}>
-
         { pageProducts && pageProducts.map((item) => (
-
           // eslint-disable-next-line dot-notation
           <Grid key={item['_id']} item sx={type === 'partners' ? gridStylesItemPartners : gridStylesItemProducts}>
             {createElement(itemComponent, { ...item })}
           </Grid>
         ))}
       </Grid>
+
       {actions}
       {(pagination && pageQty > 1) && (
       <AppPagination
@@ -164,10 +115,7 @@ const ListItems = ({ title, items, itemComponent, actions, pagination, anchor, t
 };
 
 ListItems.propTypes = {
-  title: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.bool,
-  ]),
+  title: PropTypes.string,
   actions: PropTypes.object,
   pagination: PropTypes.bool,
   anchor: PropTypes.object,
